image/requirements.txt
boto3
squareup
Pillow
qrcode
ensta
GitPython
Jinja2
jinja2-time
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
playwright